Esse commit traz uma forma simple de lidar com vários clientes ao mesmo tempo. Além de consertar um bug com sendfile no lado do cliente.

+++ include/commands.h
@@ -28,7 +28,7 @@ struct command_options
     int bitfield;
     bool asc; // se asc for falso então desc é verdade
     bool name; // se name for false então size é verdade
-    size_t file_size; // só é usado no comando send
+    int32_t file_size; // só é usado no comando send
 };
 struct status;
